decouple mlt sampler from mmlt

+++ src/integrator/bdpt.rs
@@ -42,7 +42,7 @@ impl Integrator for Bdpt {
             light_path: vec![],
         });
         parallel_for(npixels, 256, |id| {
-            let mut sampler = PCGSampler { rng: PCG::new(id) };
+            let mut sampler = PCGSampler { rng: PCG::new(id as u64) };
             let x = (id as u32) % scene.camera.resolution().x;
             let y = (id as u32) / scene.camera.resolution().x;
             let pixel = uvec2(x, y);
